requirements.txt
numpy
scipy
matplotlib
ipython
Pillow